backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "PreLegal backend API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "litellm>=1.83.9",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.39.0",
]

[tool.uv]
package = false
